tests/wpt/tests/svg/animations/animateMotion-base.html

<!doctype html>
<head>
<base href="/">
</head>
<html>
<title>Test for checking position of the svg element when there is a <base> attribute</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/SVGAnimationTestCase-testharness.js"></script>

<svg>
  <line id="line" x2="100" display="none"/>
  <rect width="50" height="50" x="0" y="0" fill="green">
    <animateMotion id="anim" begin="0s" dur="4s" fill="freeze">
      <mpath href="#line"/>
    </animateMotion>
  </rect>
</svg>

<script>
var rootSVGElement = document.querySelector("svg");
var epsilon = 1.0;

// Setup animation test
function sample1() {
    assert_approx_equals(rootSVGElement.getBBox().x, 0, epsilon);
}

function sample2() {
    assert_approx_equals(rootSVGElement.getBBox().x, 50, epsilon);
}

function sample3() {
    assert_approx_equals(rootSVGElement.getBBox().x, 100, epsilon);
}

smil_async_test((t) => {
    var rects = rootSVGElement.ownerDocument.getElementsByTagName("rect");
    rect1 = rects[0];

    const expectedValues = [
        // [animationId, time, sampleCallback]
        ["anim", 0.0,   sample1],
        ["anim", 2.0,   sample2],
        ["anim", 4.0,   sample3]
    ];

    runAnimationTest(t, expectedValues);
});

window.animationStartsImmediately = true;

</script>
