third_party/blink/web_tests/external/wpt/svg/animations/animateMotion-base-expected.txt

This is a testharness.js-based test.
[FAIL] Test for checking position of the svg element when there is a <base> attribute
  assert_approx_equals: expected 50 +/- 1 but got 0
Harness: the test ran to completion.
